+++ app/src/main/java/com/example/myapplication/ui/home/IClickListener.kt
@@ -2,7 +2,7 @@ package com.example.myapplication.ui.home
 
 interface IClickListener {
 
-    fun onItemClicked(id: Long)
+    fun onItemClicked(id: Int)
 
-    fun onItemSelected(position: Long)
+    fun onItemSelected(position: Int)
 }
